changed abstract class to interface
